Updated requirement

+++ src/main/java/com/hackdays/aws/transformer/road/repositories/StreetConfidenceRepository.java
@@ -3,5 +3,5 @@ package com.hackdays.aws.transformer.road.repositories;
 import com.hackdays.aws.transformer.road.domains.StreetConfidence;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StreetConfidenceRepository extends JpaRepository<StreetConfidence, Long> {
+public interface StreetConfidenceRepository extends JpaRepository<StreetConfidence, Integer> {
 }
